feat(ab,eap)!: Change schema resolution root to schemas dir
This frees users from having to install the schemas in the same deeply nested structure as in the reference build environment.

Details
=======

`crates/eap/src/lib.rs`:
- `DEFAULT_ACAP_SDK_LOCATION` is removed to reduce the coupling between the library and the reference build environment; the library should work in that environment but I don't think it should describe it and users may have their own idea of what "default" means (I started on this change because I thought it would be neat to infer the location from something like `SDKTARGETSYSROOT` in `acap-build` but ended up realizing that would bring no benefits in the reference environment and force all users to set that path as an indirect way of locating the schemas).
